pyproject.toml
[tool.poetry]
name = "Dashboard"
version = "0.2.0"
description = ""
authors = ["FEFU Analytics sector", "FEFU Admission"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"
fastapi = "^0.78.0"
aiofiles = "^0.8.0"
requests = "^2.28.0"
fastapi-utils = "^0.2.1"
aiohttp = "^3.8.1"
starlette = "^0.19.1"
Levenshtein = "^0.21.1"
pytz = '^2022.1'
rapidfuzz = '3.0.0'
python-dotenv = '1.0.0'

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
